% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_census.R
\name{economic_census}
\alias{economic_census}
\title{Download Economic Census Data Tables from INEGI}
\usage{
economic_census(year = NA, state = NA)
}
\arguments{
\item{year}{Year of selection}

\item{state}{Federal entitie of selection}
}
\value{
A data frame of complete Mexican economic census
}
\description{
This function download a file as a dataframe. It's a table from economic census data
from INEGI page
}
\examples{
 economic_census(year=2019, state="ags")
}
\author{
Marco Antonio Perez-Mendez
}
