% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic_units.R
\name{economic_units}
\alias{economic_units}
\title{Economic units}
\usage{
economic_units(data = NA)
}
\arguments{
\item{data}{data downloaded with economic_census() function}
}
\value{
a data.frame
}
\description{
Economic units
}
