% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbt_munR.R
\name{pbt_mun}
\alias{pbt_mun}
\title{Filter Municipal data from Economic Census}
\usage{
pbt_mun(data = NA)
}
\arguments{
\item{data}{is a filtered data}
}
\value{
A data.frame of the infile
}
\description{
This function filter municipal gross total product data from economic census
}
