% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM_04_19.R
\name{PIBM_04_19}
\alias{PIBM_04_19}
\title{PBT Municipal 2004-2019
Join municipal PBT data for several years}
\usage{
PIBM_04_19(state = NA)
}
\arguments{
\item{state}{name of federal entitie}
}
\value{
A data.frame of the infile
}
\description{
This function joins municipal total gross product for several years and
all federal entities
}
